# server/public/api/openapie.yaml
openapi: 3.1.0
info:
  title: CTU eLibrary API
  version: 1.0.0
  description: API quản lý thư viện điện tử của Đại học Cần Thơ
  x-fiddle-import-file: true
servers:
  - url: /api
    description: API Server
paths:
...
  /readers/holds:
    get:
      operationId: getHolds
      tags:
        - Readers
      summary: Get reader hold requests
      description: Xem danh sách yêu cầu giữ sách của độc giả hiện tại
      security:
        - accessToken: []
      responses:
        "200":
          description: Danh sách yêu cầu giữ sách
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HoldRequest"
                title: GetReaderHoldsOk
        "401":
          description: Không có quyền truy cập
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      parameters: []
components:
  securitySchemes:
    accessToken:
      type: apiKey
      name: accessToken
      in: cookie
      description: ""
  schemas:
...
    HoldRequest:
      type: object
      properties:
        id:
          type: string
        reader:
          type: string
        book:
          type: string
        holdDate:
          type: string
          format: date-time
        expiryDate:
          type: string
          format: date-time
      required:
        - reader
        - book
        - holdDate
        - expiryDate
        - id
      description: ""
...
    Error:
      type: object
      properties:
        message:
          type: string
        error:
          type: string
...